+++ libs/ui/svelte/components/Wallpaper/components/ThemedWallpaper.svelte
@@ -1,21 +1,44 @@
 <script lang="ts">
-  import { onMount } from "svelte";
   import BackgroundByLayers from "../../BackgroundByLayers/BackgroundByLayers.svelte";
   import type { BaseProps } from "../types";
   import { getWallpaperStyles } from "../utils";
+  import { players } from "../state.svelte";
+  import { convertFileSrc } from "@tauri-apps/api/core";
 
   let { definition, config, onLoad }: Pick<BaseProps, "definition" | "config" | "onLoad"> =
     $props();
 
-  onMount(() => {
-    onLoad?.();
+  const player = $derived(players.value.find((p) => p.default));
+  const thumbnailSrc = $derived(player?.thumbnail ? convertFileSrc(player.thumbnail) : null);
+
+  $effect(() => {
+    if (onLoad) {
+      if (!definition || definition.type !== "MediaPlayer" || !thumbnailSrc) {
+        onLoad();
+      } else {
+        fetch(thumbnailSrc).finally(() => onLoad());
+      }
+    }
+  });
+
+  $effect(() => {
+    if (thumbnailSrc) {
+      document.documentElement.style.setProperty(
+        "--media-player-thumbnail",
+        `url('${thumbnailSrc}')`,
+      );
+    }
   });
 </script>
 
-{#if !definition || !config}
-  <BackgroundByLayers class={["wallpaper", "default-wallpaper"]} />
+{#if !definition || (definition.type === "MediaPlayer" && !player)}
+  <BackgroundByLayers id="@default/wallpaper" class={["wallpaper", "default-wallpaper"]} />
 {:else}
-  <BackgroundByLayers id={definition.id} class="wallpaper" style={getWallpaperStyles(config)}>
+  <BackgroundByLayers
+    id={definition.id}
+    class="wallpaper"
+    style={config ? getWallpaperStyles(config) : undefined}
+  >
     {@html `<style>@scope { ${definition.css || ""} }</style>`}
   </BackgroundByLayers>
 {/if}