+++ libs/ui/svelte/components/Wallpaper/state.svelte.ts
@@ -1,0 +1,6 @@
+import { invoke, SeelenCommand, SeelenEvent, subscribe } from "@seelen-ui/lib";
+import { lazyRune } from "../../utils";
+
+export const players = lazyRune(() => invoke(SeelenCommand.GetMediaSessions));
+subscribe(SeelenEvent.MediaSessions, players.setByPayload);
+await players.init();